Enable local cache (`--cache-reuse`) by default
This makes `--cache-reuse` enabled by default. As same as `buildg dap`, this
commit introduces a lock on the BuildKit root not to conflict among buildg
processes.
If the 2nd buildg process launched, it'll print a warning and use a temporary
rootdir location.
In the future version, parallel invocation of buildg should share the cache.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>